now it works. The new and old probe updates are different somewhere...
