+++ templates/simulation_test_OPR.py
@@ -15,7 +15,7 @@ p.io.home = "/tmp/ptypy/"
 p.io.autosave = u.Param()
 p.io.autosave.interval = 20
 p.io.autoplot = u.Param()
-p.io.autoplot.active = False
+p.io.autoplot.active = True
 p.io.interaction = u.Param()
 
 p.scans = u.Param()
@@ -57,30 +57,30 @@ p.scans.MF.illumination.aperture.edge = 10
 
 
 p.engines = u.Param()
-p.engines.engine00 = u.Param()
-p.engines.engine00.name = 'DM'
-p.engines.engine00.numiter = 3
-p.engines.engine00.numiter_contiguous = 5
-p.engines.engine00.overlap_max_iterations = 2
-p.engines.engine00.fourier_relax_factor = 0.01
+# p.engines.engine00 = u.Param()
+# p.engines.engine00.name = 'DM'
+# p.engines.engine00.numiter = 3
+# p.engines.engine00.numiter_contiguous = 5
+# p.engines.engine00.overlap_max_iterations = 2
+# p.engines.engine00.fourier_relax_factor = 0.01
 
 p.engines.engine01 = u.Param()
 p.engines.engine01.name = 'DMOPR'
-p.engines.engine01.numiter = 1
-p.engines.engine01.numiter_contiguous = 5
+p.engines.engine01.numiter = 100
+p.engines.engine01.numiter_contiguous = 1
 p.engines.engine01.overlap_max_iterations = 2
 p.engines.engine01.fourier_relax_factor = 0.01
 p.engines.engine01.IP_metric = 1.
 p.engines.engine01.subspace_dim = 10
-
-p.engines.engine02 = u.Param()
-p.engines.engine02.name = 'MLOPR'
-p.engines.engine02.numiter = 100
-p.engines.engine02.numiter_contiguous = 5
-p.engines.engine02.reg_del2 = True
-p.engines.engine02.reg_del2_amplitude = 1.0
-p.engines.engine02.scale_precond = True
-p.engines.engine02.subspace_dim = 10
+#
+# p.engines.engine02 = u.Param()
+# p.engines.engine02.name = 'MLOPR'
+# p.engines.engine02.numiter = 100
+# p.engines.engine02.numiter_contiguous = 5
+# p.engines.engine02.reg_del2 = True
+# p.engines.engine02.reg_del2_amplitude = 1.0
+# p.engines.engine02.scale_precond = True
+# p.engines.engine02.subspace_dim = 10
 
 P = Ptycho(p, level=4)
